compiles and builds the database, but needs to be refactored.
The problem here is that we don't have access to the db where we want
it and that may be a big deal.  Having to have the db in a future
makes me think we may actually need to transfer it into the future and
then take it out again.  I guess that's why I'm prototyping...