This branch is probably useless.
It compiles, but has no functionality and tons of code commented out.
I was disappointed in my rexie version not working under Firefox, so I
started to make a version that used only web-sys, but I got curious
about index-db and made a version that used index-db and it works with
Firefox, so my inclination is to use index-db, because I like how it
treats transactions.

Turns out, after I made a version with index-db, I found that removing
key_path in my rexie version was all that was needed to work with
Firefox.

So, I don't think there's any reason for me to make a version that
works just using web-sys, but perhaps I'll change my mind.